refactor/cleanup - reduce contents of event-track.js - move helper functions to appropriate util files
